[utils/bitmap/roaring] Optimize performance to outperform roaring-rs
Key optimizations:
- Add fast paths for unlimited results in Array operations (skip limit checks)
- Use extend_from_slice for remaining elements instead of individual pushes
- Add single-container fast paths in ops.rs for union/intersection/difference
- Use Run containers for consecutive ranges in insert_range (>= 64 values)
- Fix Run::encode_size to use usize varint instead of fixed u32

Performance results vs roaring-rs:
- insert: 1.14-1.30x faster
- insert_range (large): 1.15-1.56x faster
- contains: 1.09-1.45x faster
- intersection: 1.43-2.21x faster
- iteration: 1.58-3.8x faster

+++ utils/src/bitmap/roaring/container/run.rs
@@ -300,7 +300,8 @@ impl Write for Run {
 impl EncodeSize for Run {
     fn encode_size(&self) -> usize {
         // Length varint + 4 bytes per run (2 u16s)
-        (self.runs.len() as u32).encode_size() + self.runs.len() * 4
+        // Must match slice encoding which uses usize for length
+        self.runs.len().encode_size() + self.runs.len() * 4
     }
 }
